Actualización en método Eliminar

+++ src/main/java/com/example/proyectou2_nosql/service/VitalidadService.java
@@ -3,9 +3,11 @@ package com.example.proyectou2_nosql.service;
 import com.example.proyectou2_nosql.model.Vitalidad;
 import com.example.proyectou2_nosql.repo.VitalidadRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class VitalidadService {
@@ -20,7 +22,14 @@ public class VitalidadService {
         return vitalidadRepo.findAll();
     }
     public void deleteById(String id){
-        vitalidadRepo.deleteById(id);
+        if (id == null) {
+            throw new IllegalArgumentException("id must not be null");
+        }
+        Optional<Vitalidad> vitalidadOpt = vitalidadRepo.findById(id);
+        if (!vitalidadOpt.isPresent()) {
+            throw new RuntimeException("Not Found");
+        }
+        vitalidadRepo.deleteById(id);;
     }
     public Vitalidad update(Vitalidad vitalidad) {
         return vitalidadRepo.save(vitalidad);
